Compile warning fix

+++ src/visSurvey/vishorizon2ddisplay.cc
@@ -336,7 +336,7 @@ bool doWork( od_int64 start, od_int64 stop, int )
 		if ( zaxt_ )
 		{
 		    const TrcKey tk( geomid, rc.col() );
-		    pos.z_ = zaxt_->transformTrc( tk, pos.z_ );
+		    pos.z_ = (double)zaxt_->transformTrc( tk, (float)pos.z_ );
 		}
 
 		if ( !mIsUdf(pos.z_) )
